Split fetch and parse logic between DictionaryResource and utils.Dictionary, and add tests
